写的时候写错了，应该用int，写成long了

+++ qmq-store/src/main/java/qunar/tc/qmq/store/SortedPullLogTable.java
@@ -325,7 +325,7 @@ public class SortedPullLogTable implements AutoCloseable {
                     buffer.writeBytes(consumerBytes);
                     buffer.writeLong(indexEntry.startOfPullLogSequence);
                     buffer.writeLong(indexEntry.baseOfMessageSequence);
-                    buffer.writeLong(indexEntry.position);
+                    buffer.writeInt(indexEntry.position);
                     ByteBuffer nioBuffer = buffer.nioBuffer();
                     Checksums.update(crc, nioBuffer, nioBuffer.limit());
                     boolean result = tablet.appendData(nioBuffer);
